hsm: Remove support for size in sectors as int
We used to support volume size either as number of sectors using int, or
as number of bytes using a string. Engine is using size in bytes since
first git commit. Hosted engine setup is also using size in bytes.  We
don't know about old code using number of sectors.

This patch removes the support for size in sectors.  To prevent
incorrect usage we raise InvalidParameterException if called with an
integer.

Change-Id: I52957eec163973fd60a95a9bf523131f3c83a9da
Bug-Url: https://bugzilla.redhat.com/1221603
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47381
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-on: https://gerrit.ovirt.org/47418
Tested-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>