guest-agent: desktopLock exception on EBADF should be ignored
There's no more logging of the EBADF exceptions on desktopLock thrown
when the sockets are closed, since this is very likely to happen on a
migration because we are receiving multiple display disconnect events
which are delayed for 2 seconds before being processed to avoid the
screen to be locked in migration cases - in allows the lock not to be
executed when we're already connected again (e.g. on a fast reconnect
of the display)
Therefore the exception on its own does not provide much value and
therefore is explicitely swallowed for the desktopLock case.

All other exceptions are still logged, in case something unexpected
happens.

Additional note on the multiple disconnect events: We have
implemented and merged a fix already to reduce this to one time.
However this depends on a fixed libvirt - because they had a bug
that they did not report the client remote ports and therefore
we couldn't make a decision whether or not the disconnect event
should be handled or not - so we handle it 4 times for now.

Change-Id: I3cd032471396a97999c2cea5c88ae04ec8ca4028
Backport-To: 3.6
Bug-URL: https://bugzilla.redhat.com/1151863
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47404
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>