betterAsyncore: Fix AttributeError on disconnect
Recently we added logging when SSLError is raised during recv(), which
happens when a client disconnects in an unclean way. Unfortunately,
logging the error raises an AttributeError and fail to close the
connection as intended:

    >>> from yajsonrpc.betterAsyncore import Dispatcher
    >>> d = Dispatcher()
    >>> d.log.error("message")
    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
    AttributeError: 'function' object has no attribute 'error'

Dispatcher.log is a method, not a logger instance.

This patches adds a "vds.dispatcher" logger and use it to log the
message.

Since this is an expected event, and logging error and warnings is also
duplicated to the system log, we use now info message.

Change-Id: I84387711167de39ea9364f0717f23ba4aa95edf5
Bug-Url: https://bugzilla.redhat.com/1272496
Backport: 3.6
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47406
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/47421
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Francesco Romani <fromani@redhat.com>