virt: Allow _recover() to call vm.cont() in any VM state
By default, cont() returns error if the VM is in one of the following
states:

vmstatus.MIGRATION_SOURCE
    Migration is in progress, VM status should not be changed till the
    migration finishes.

vmstatus.SAVING_STATE
    Hibernation is in progress, VM status should not be changed till the
    hibernation finishes.

vmstatus.DOWN
    VM is down, continuing is not possible from this state.

But if an error occured when performing an operation on a VM and
SourceThread._recover() is called, it must be able to call vm.cont() for
recovery purposes regardless of the VM's status.

To implement this, ignoreStatus parameter was added to vm.cont(). When
set to True, it forces vm.cont() to run even in the states listed above.

Change-Id: I5b1c7b4eecacf87ece48dc563fd2da294af0510b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1238536
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47527
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>