guestagent: Add support for containers message
This patch introduces the support of reporting the list of containers,
sent by the guest agent and exposing it to the VDSM API.

Sent by the guest agent is a list of active containers running on the guest
virtual machine.
Each container includes the id, names/aliases, image name, ran command and
currently reported status.
This information will be forwarded as is to the API.

Change-Id: If42efd4d2743e231a5b012c2901b8cf372812898
Guest-Agent-API-Version: 3
Bug-Url: https://bugzilla.redhat.com/1273399
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47519
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>