DynamicBridge: Log TypeError exceptions raised during dispatch
When dispatching via the DynamicBridge, the dispatched function may
raise a TypeError (typically due to a coding error).  Currently we
swallow the TypeError and raise InvalidCall instead.  This hides the
underlying problem making it harder to fix.  Keep the current behavior
but log the original exception too.

Change-Id: I27181f94a39fc77eb60f1464c06d173c38e19367
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47546
Reviewed-by: Greg Padgett <gpadgett@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>