setupNetworks: moving nics between bonds
This patch fixes a problem where a slave removed from
one bond could sometimes not be assigned to another
bond in the same request.
This happened because removing, editing and adding of
slaves to a bond was done bond by bond, so if a nic
was added to a bond which was handled before the bond
from which the nic was removed the operation failed.
Example:
bond0: nic0, nic1   => nic2, nic3
bond1: nic2, nic3   => nic0, nic1
In this situation first bond0 is processed: nic0 and
nic1 will be removed, but nic2 and nic3 can not yet
be removed as they are part of bond1 (validation fails).
Processing of bond1 will introduce errors, as nic2 and
nic3 will be removed (already assigned to bond0),
which will in effect reset their ifcfg files, and cause
the previous attachement to bond0 to be undone.

The patch changes the validation to check if a slave
added to a bond is deleted from another bond.
Also the removing of slaves from all bonds is done
before any other bond operation.

Bug-Url: https://bugzilla.redhat.com/1269175
Change-Id: Ia43fbfe90d1335ed4a740a1bc4f09610da2ecdce
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50596
Continuous-Integration: Jenkins CI
Reviewed-by: Edward Haas <edwardh@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>