hotunplug: Supporting lun disks fix.
While hotunplugging, VDSM queries libvirt for the unplugged disk to
verify it was removed. This check is in _isDriveAttached, which verifies
that the disk's serial is no longer among the VM's disks. As LUN disks
currently do not have a serial specified in VDSM flows,
after unplugging, _isDriveAttached queries for empty disk serial to
detect whether the disk is still there, and returns true in case there
are disks with empty serials- such as CD-ROM disks.

This patch fixes _isDriveAttached to check the disk's presence by
verifying that the disk's path no longer exists in the xml, rather
than the serial- to make it return the right value when unplugging
LUN disks.

[Verification of the patch]:
- For each disk type in (file disk, block disk, direct LUN disk)
  - Try unplugging from a VM with OS installed- should work. [verified]
  - Try unplugging from a VM without OS installed- should 
    fail with timeout, since libvirt keeps the disks in its XML 
    description of the VM. [verified]

Change-Id: I826eba42903167988da1d02b916feb8fcd19258e
Bug-Url: https://bugzilla.redhat.com/1270834
Signed-off-by: Amit Aviram <aaviram@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47795
Continuous-Integration: Jenkins CI
Reviewed-by: Daniel Erez <derez@redhat.com>
Tested-by: Daniel Erez <derez@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47846
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47926
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Continuous-Integration: Yaniv Bronhaim <ybronhei@redhat.com>