lib: taskset: commands should run on all cpus
The current affinity reset logic has a bug in the
case the sysadmin puts offline one or more cpu.
This is because we use the range notation, with
the upper limit given by sysconf('SC_NPROCESSORS_ONLN')-1

Let's say we have 16 cpus, and sysadmin puts offline
cpu #11. Now we will have 15 online cpus, so we'll
build our range as 0-14, which excludes the perfectly
fine cpu #15.
Note that the range still includes the offline CPU, but
it is safe to trust the kernel NOT to use it.
However, this buggy logic make VDSM rule out TWO cpus,
one implicitely (included in range, but offline), plus
the one erroneously excluded by the range.

To fix this, we simply always use SC_NPROCESSORS_CONF
to express the maximum physical range, and we let the
kernel handle the holes on this range.

Change-Id: I0d17ab063356bcfde1192d545a7d41114eeab0be
Backport-To: 3.6
Backport-To: 3.5
Bug-Url: https://bugzilla.redhat.com/1279431
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47801
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48786