vdscli: map invocation params to dictionary
We need to be more flexible with parameters provided during execution of
vdscli methods. Now we use api schema to name all the params so we can
match them properly on the server side.

With this change we can call vdscli methods as regular method and do
mapping behind the scene.

Code used for xmlrpc looked like:

self.vdscli.addNetwork(bridge,
                       *self._get_net_args(vlan, bond, nics, opts))

In above case we are required to prepare named params as array.

Now we can call it:

self.vdscli.addNetwork(bridge,
                       vlan=vlan, bond=bond, nics=nics, opts=opts)

as well as we can provide only optional params as we want.

This is first step which enable us to provide client side schema
validation.

It is required to extract schema files from jsonrpc rpm due to
dependency needed for python rpm. Now we make schema rpm required for
jsonrpc and python rpm.

Change-Id: Ibac6eab3c519becb29d2b3551111d671bbb79df5
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/45429
Reviewed-by: Yeela Kaplan <ykaplan@redhat.com>
Tested-by: Yeela Kaplan <ykaplan@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>