json cli: use ssl when creating a socket
We want to be compatible with xmlrpc cli using regular ssl,
instead of default m2crypto.

During handshake, m2crypto is trying to compare the hostname
given to the certificate hostname which might not
always be compatible.
Default ssl is doing the handshake without it,
helping us avoid this incompatibility with
no harm to performance.

Change-Id: I2fa15584eed90945cdf2eb8cad17ed7697678932
Signed-off-by: Yeela Kaplan <ykaplan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/45611
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>