tests: Add FakeLVM LV properties tests
Building on the example environment in FakeLVMSimpleVGTests we create a
single logical volume.  We then verify that the properties and
attributes of that LV match what is returned by the real lvm module.
Some problems were found in FakeLVM and are resolved here.

Since createLV accepts the size in MB, two of the manifest_tests needed
to be fixed.

Change-Id: I1bca30db49db84fc5d42ff8736a4c69f17b17247
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47859
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>