sampling: rename attributes for clarity
VMBulkSampler has one ExpiringCache to track VMs/Domains believed
to be stuck, in order to avoid further clogging.
Due to an old mistake, the 'ttl' of this cache was set using a variable
named 'timeout'. The choice of names is unfortunate and source of
confusion.

This patch fixes that.

Change-Id: Ida543cf6f41e6e7a0b99277eb447117115fb4df1
Rename-Only: yes
Backport-To: 3.6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47893
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>