storagetestlib: use a real metadata file for FileSD manifests
When building a fake FileStorageDomainManifest place the metadata in
the right place within the domain's file hierarchy (as opposed to
keeping it in memory).  This is needed since code in VolumeMetadata
sometimes instantiates its controlling manifest and the metadata must be
present in the proper place when testing this code.

Change-Id: I1e09b3e2199c638452214de4c2e4dfd826b51242
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48695
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>