lib: executor: log task name when discarded
To improve the debuggability of the executor, we
add the name of the task which was stuck/slow
when a worker is discarded.

Doing so it is easier to understand of what's going on
and why workers are being replaced, instead of doing
backtracking in the logs.

Change-Id: I40584ee9a99839a4cb56d2f4d908d274f3a46638
Bug-Url: https://bugzilla.redhat.com/1250839
Backport-To: 3.6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48191
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48415