vdsm.conf: Add drop-in dir
A drop-in dir allows the user to put partial configurations
into a drop-in directory.
This will allow to either set the configuration or override
already set configurations using the partical configuration files.

To use this feature:

$ mkdir /etc/vdsm/vdsm.conf.d/
$ echo -e "[vars]\nfake_kvm_support = true\n" > \
  /etc/vdsm/vdsm.conf.d/50-fake-kvm.conf

The semantics of the directories and the search order is
as follows.

- /etc/vdsm/vdsm.conf - for user configuration. We install this
  file if missing, and never touch this file during upgrade.
- /etc/vdsm/vdsm.conf.d/ - for admin drop-in conf files.
- /usr/lib/vdsm/vdsm.conf.d/ - for vendor drop-in configuration files.
- /var/run/vdsm/vdsm.conf.d/ - for admin temporary configuration.

Files with a .conf suffix can be placed into any of the
vdsm.conf.d drop-in directories.

Change-Id: I707a1865e8d60dc4dcdc0e681b52c07c75f1c409
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1279555
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>
Signed-off-by: Irit Goihman <igoihman@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48317
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Tested-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>