storage: perform image deletion in task
deleteImage() always schedules a task after it's synchronous part, the
task is currently used to wipe the image (if requested) and in case of
deletion without wipe it performs nothing.

The synchronous part of deleteImage() renames the image to clarify that
it's about to be deleted by adding a prefix and then deletes the image.
On storage under heavy load any of those operations can be slow, as the
engine has a default timeout of 180s for vdsm operation on some cases we
might fail on timeout. In order to lower the chance to get that timeout
this patch moves the image deletion to occur always on the task (same as
happens today when wiping the image).

The assumption that the engine relies on currently is that when a call to
deleteImage() returns the image is no longer available and that it's id
is "free" to use (whether it was deleted on the synchronous part or a
wipe is ongoing), as we still rename the image in the synchronous part
this assumption is still valid and the engine shouldn't be affected by
it.

This patch change deleteImage() to perform only the image rename, while
purgeImage() will be used to free the space taken by the image.

Change-Id: I1e61d38439192d193f3f806e08661b61a65e7836
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1270220
Signed-off-by: Liron Aravot <laravot@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48477
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Freddy Rolland <frolland@redhat.com>