issci: Fix equality and hash when username or password are None
When fetching iscsi session info, username of password can be None, but
iscsi.ChapCredentials was not handling these values correctly in __eq__
and __hash__.  In particular, we used to compare password.value and
include it in the hash. This would raise AttributeError when password is
None.

Now all combinations are handled properly; we compare password objects,
using ProtectedPassword.__eq__, and include hash(password), using
ProtectedPassword.__hash__.

This issue was hidden because practically username is None when password
is None, and the equality test fail before we try to compare None.value.
The hash issue is hidden because we never use __hash__ for
ChapCredentials. So this is mainly a correctness fix that should not
effect current code, but it may be needed for fixing the related bug.

Change-Id: I95bda5d07506ecbfb76ec5397717fe05b53f97ec
Bug-Url: https://bugzilla.redhat.com/1279485
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48480
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Ala Hino <ahino@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50939
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>