iscsi: Fix credentials initialization when not using CHAP
There was a mismatch between the way we initialize ChapCredentials when
receiving empty credentials from engine, and when reading empty
credentials from active session.

This was broken when we started to protect password with
ProtectedPassword.  When getting a request from engine we used
ProtectedPassword(""), while the iscsi code used None. This causes
session disconnect to fail with "cred mismatch".

Now we have a note explaining the dependency and we use the same way to
initialize the credentials:

- When receiving request from engine, empty username or password are
  stored as None.
- When reading requests from active session, empty username or password
  are stored as None, including the special empty values "<NULL>" and
  "(null)".

Code was moved so username and password normalization is near
ChapCredentials initialization in both flows.

Change-Id: I65e1177bc33998b8eba34b0fd04ec62cf8bf1a2c
Bug-Url: https://bugzilla.redhat.com/1279485
Reported-By: Elad Ben Aharon <ebenahar@redhat.com>
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48483
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/50940
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>