mkimage: setup right permissions before mkisofs
We don't want to depend on the 'childUmask' argument of cpopen anymore,
so that argument could be dropped in future release, to make the
interface of cpopen identical to the one of the standard subprocess
module.

The 'childUmask' argument was added to make sure that the iso image we
create has the right permissions right from the file creation, to avoid
possible information leak.

We can reach the same goal by preparing one hollow file with the right
attributes on the same path mkisofs will use. mkisofs will truncate the
(already empty) content of the file we prepared, and keep the
other attributes.

There are no planned change in functionality.

Change-Id: Ibc971a39ae2a8eaf7af934bb0922647e4676d03f
Bug-Url: https://bugzilla.redhat.com/1278414
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48538
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI