hsm : Do not resize on disconnectStorageServer
When a host is moved to maintenance state, it will
disconnect from connected storage servers.

Then, sdCache.refreshStorage() is called. It will rescan
the devices, and attempt to resize devices if needed.

However, in this flow, resizing devices is not needed and may
cause exceptions when trying to access sysfs when the iscsi
connections are being removed, and needlessly delay the
maintenance operation.

Change-Id: I78e236a851b7a0e9ab8f3c86e180cc50fde4ad56
Bug-Url: https://bugzilla.redhat.com/1281909
Signed-off-by: Fred Rolland <frolland@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48578
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI