vm: Remove empty RNG devices left after 3.5->3.6 upgrade
Clean out unknown devices that are created for RNG devices by VDSM 3.5
and are left in the configuration after upgrade to 3.6 until the VM is
rebooted. These devices do not contain 'source' spec param so they
cause exceptions on newer VDSM versions that expect 'source' to be
present.

Change-Id: I2580dcb335970a4bb5c4c87fe6ea9fcd8e494df5
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1282581
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48621
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI