net: break netinfo-netconfpersistence dependency
netinfo module currently depends on netconfpersistence module
because the use of RunningConfig during for the accurate reporting
of dhcp from VDSM. This prevents code that lies in
netconfpersistence to use netinfo because a circular dependency is
created. Because of that, KernelConfig has to abuse netinfo.NetInfo
object in order to avoid importing its parent module (netinfo).
This patch separates KernelConfig from netconfpersistence, thus
breakiing the circle and allows a simplification of both KernelConfig
and netinfo.NetInfo in the following patches.

Change-Id: Ie64081f13b249ff7a7595ae56d00d1b218428d9c
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48824
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>