debuging: Use __repr__ instead of __str__
Recently we are implementing __str__ for objects, for improved logging:

    logging.debug("Discarding task %s", self.task)

We are using the style used by Python:

    >>> threading.RLock()
    <_RLock owner=None count=0>

Implementing __str__ is fine for logging, but when using the Python
interactive shell (e.g, manhole), the output is not very useful:

    >>> class Foo(object):
    ...     def __str__(self):
    ...             return "<Foo bar=2 at 0x%x>" % id(self)
    ...
    >>> f = Foo()
    >>> f
    <__main__.Foo object at 0x7f5f85441c90>

For debugging, we should implement __repr__ instead of __str__.
According to the language reference[1], when using str(), if object does
not implement __str__, its __repr__ is used.

This patch replaces most relevant __str__ methods with __repr__. Some
methods needs more work and left for later.

One instance implementing __repr__ that return the same output as
object.__repr__ was removed.

[1] https://docs.python.org/2/reference/datamodel.html#object.__repr__

Change-Id: I2fd724e32dfde1ca484343730f2e55c37a59fdf3
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48865
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>