exception: Add exceptions for vdsm.define errors
Add an exception class for each error in vdsm.define.

The long term goal is simplifying error handling, so code can raise
error in whatever layer the error happened, and single error handler
will generate the response using the error response() method.

As a temporary step, define.errCode is using the new exceptions to build
the response dict. When callers are updated to raise the new exceptions,
we can drop the errCode dict and the define module.

Change-Id: I01b736c06414a3af758ad6bdabddb6c25b620756
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48871
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>