stomp: Send all the pending messages in handle_write
We used to send one pending message in handle_write, and poll again,
which would wake up the reactor immediately since the socket is still
writable.

When socket is writable, we like to send all the pending data until
there is no data to send, or the socket buffers are full.

This change should send data quicker and consume less cpu time.

Change-Id: I35dad03942cb53f95c57cc6b48bd3ae93885a0d2
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48888
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>