net: Flatten KernelConfig logic
KernelConfig, which can now use the netinfo module directly can let
go of most of its logic and make it as module level.
This makes the bypass @staticmethods of NetInfo redundant.

After the netinfo API is refactored, we should decide if we need the
dependency of KernelConfig in a NetInfo object. The only real
dependency is on NetInfo.getNicsVlanAndBondingForNetwork and on
NetInfo.Bondings.

Change-Id: I5e0df1b4355cafb9f4d3a6da070b9384b858aba0
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48893
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>