hsm.py: Deprecate getVGList
oVirt Engine commit 4312bfd removed all the usages of getVGList from the
engine's code.

Since this verb is useless to us (the same functionality is covered by
getVGInfo and getDeviceList), and since it's no longer called, this
patch flags it as @deprecated, with the intent to remove it completely
in the future.

Change-Id: Icac1e9f317af4f6089e11d647b9df361bd627889
Signed-off-by: Allon Mureinik <amureini@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48998
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Adam Litke <alitke@redhat.com>