stomp: make sure that stomp connection is active before sending messages
In may happen that client sends SEND frame before CONNECT and SUBSCRIBE
frames are processed by the server which can lead to server having no
knowledge to whom send response because subscription is not finalized
yet. We need to wait for server to respond with CONNECTED frame before
sending any rpc requests.

Here is the flow how the issue happens:
1. Client connects to the server
2. Client sends CONNECT and SUBSCRIBE frames
3. Client sends SEND fame
4. Server receives all 3 frames and starts to process them
5. Server attempts to send MESSAGE frame
6. Server sends CONNECTED


Change-Id: I4f6881d7966fa47031d027740633ca83c834387f
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>