net: Remove reporting of bond custom options from RunningConfig
This reporting is ignored in engine, is not symmetric to network
custom properties as VDSM does not report such a field on a network
and also deepens the reliance of netinfo on RunningConfig.
Since we want to remove this dependency soon, this is one step in
this direction.

Change-Id: I5863d67b2b2b9ce15cee9c447983d3ba2e0f23ce
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/49081
Continuous-Integration: Jenkins CI
Reviewed-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>