guestagent: Send complete messages
When forwarding message to the guest agent, we used socket.send(), which
may write only part of the message and return the number bytes written.

This may lead to corrupted messages, confusing the receiver and failing
both the current request and the next one, which will be appended to the
partial request on the receiver side.

Now we use socket.sendall(), returning when all bytes were sent.

Note that this is not a complete solution, since we are working with
non-blocking socket, and it may fail with EAGAIN if the socket buffers
are full. In this case we should poll the socket until it is writable
and send the rest of the data. This is not handled in the current patch.

Change-Id: Ic832634c4d670aa88fe79a50b2e57246d08b4a08
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/49339
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>