monitor: Do not release host id during shutdown
When shutting down, all domain monitors are stopped and release acquired
host ids. This will cause sanlock to terminate processes holding a
resource on these domains, such as the qemu process for the hosted
engine vm.

This behavior was fine when the only resource held is the spm resource,
but now we must support the hosted engine vm, that must not be killed
when vdsm terminates. In the future we may like to use sanlock resources
for other vms as well.

Now we detect when a monitor is stopped during a vdsm shutdown, and do
not release the host id.

This change will effect users trying upgrade vdsm without putting a host
to maintenance before the upgrade. This was never supported but users
could get away with this by stopping vdsm. Now sanlock will refuse to
stop after vdsm is stopped, and killing sanlock will cause the watchdog
to reboot the machine.

Change-Id: Ib14fa48abf9c98fdce268c4675d71aab05c92b10
Bug-Url: https://bugzilla.redhat.com/1269768
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/49386
Reviewed-by: Sandro Bonazzola <sbonazzo@redhat.com>
Reviewed-by: Martin Sivák <msivak@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Tested-by: Roy Golan <rgolan@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-on: https://gerrit.ovirt.org/49596
Reviewed-by: Roy Golan <rgolan@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>