sp: Allow attaching of monitored domain
When attaching a storage domain to the pool, the storage domain may be
an external domain being monitored via startMonitoringDomain, such as the
hosted engine storage domain.

In this case, the monitor may have already acquired the host id, or will
acquire it during the attach operation. We cannot tell since monitors
are using asynchronous acquire.

We must not release the host id when the domain is attached, since this
will cause sanlock to kill any process holding a resource on this
storage domain, such as the qemu process running the hosted engine vm.

We had a magical, sneaky, and unclear code for attaching a monitored
domain when starting a monitor twice. This code is now simplified and
documented, and logging is improved. This will be replaced with explicit
attach later.

Change-Id: Ia2fc1b19a6e70d4001a8f6d8c14bae7f65d0c192
Bug-Url: https://bugzilla.redhat.com/1269768
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48217
Continuous-Integration: Jenkins CI
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>